refactor: move formatByes from common to Number helper
